src/AI/requirements.txt
twilio
openai
Flask
requests
python-dotenv
